split generic checks
