+++ qa/qa/checks/generic/flagged_terms.py
@@ -1,0 +1,69 @@
+"""Generic flagged-word checks."""
+
+from qa.checks.base import BaseGenericPatternCheck
+
+
+class NoAnnotationSymbols(BaseGenericPatternCheck):
+    name = "no_annotation_symbols"
+    display_name = "No Annotation Symbols"
+    id = 10015
+    version = "1.0.0"
+    description = "The value does not contain invalid characters such as *, #, ^, or @."
+    failure_message = "Unusual character detected."
+
+    _pattern = r"\*|#|[^\\]\^|@"
+
+
+class DoesNotContainAnonymous(BaseGenericPatternCheck):
+    name = "does_not_contain_anonymous"
+    display_name = "Does Not Contain Anonymous"
+    id = 10019
+    version = "1.0.0"
+    description = "The value does not contain the word 'anonymous'."
+    failure_message = "Contains 'anonymous'."
+
+    _pattern = r"(?i)anonymous"
+
+
+class DoesNotContainCorresponding(BaseGenericPatternCheck):
+    name = "does_not_contain_corresponding"
+    display_name = "Does Not Contain Corresponding"
+    id = 10020
+    version = "1.0.0"
+    description = "The value does not contain the word 'corresponding'."
+    failure_message = "Contains 'corresponding'."
+
+    _pattern = r"(?i)corresponding"
+
+
+class DoesNotContainAccepted(BaseGenericPatternCheck):
+    name = "does_not_contain_accepted"
+    display_name = "Does Not Contain Accepted"
+    id = 10041
+    version = "1.0.0"
+    description = "The value does not contain the word 'accepted'."
+    failure_message = "Contains 'accepted'."
+
+    _pattern = r"(?i)accepted"
+
+
+class DoesNotContainSubmitted(BaseGenericPatternCheck):
+    name = "does_not_contain_submitted"
+    display_name = "Does Not Contain Submitted"
+    id = 10042
+    version = "1.0.0"
+    description = "The value does not contain the word 'submitted'."
+    failure_message = "Contains 'submitted'."
+
+    _pattern = r"(?i)submitted"
+
+
+class DoesNotContainEtAlWithPeriod(BaseGenericPatternCheck):
+    name = "does_not_contain_et_al_with_period"
+    display_name = "Does Not Contain Et Al With Period"
+    id = 10059
+    version = "1.0.0"
+    description = "The value does not contain the malformed abbreviation 'et. al.' (a period should not follow 'et')."
+    failure_message = "Contains 'et. al.'."
+
+    _pattern = r"(?i)\bet\.\s*al\b"
