+++ qa/qa/checks/generic/length.py
@@ -1,0 +1,76 @@
+"""Generic length checks."""
+
+from qa.checks.models import Result, Offset, OnFailurePolicy, QaDataRegistry
+from qa.checks.base import BaseGenericCheck
+
+
+class NotTooLong(BaseGenericCheck):
+    name = "not_too_long"
+    display_name = "Not Too Long"
+    id = 10036
+    version = "1.0.0"
+    description = "The value does not exceed the maximum character length."
+    failure_message = "Too long."
+
+    def __init__(
+        self,
+        *,
+        max_chars: int,
+        on_failure_policy: OnFailurePolicy,
+        data: str,
+        field: str,
+    ) -> None:
+        super().__init__(on_failure_policy=on_failure_policy, data=data, field=field)
+        self.max_chars = max_chars
+
+    @property
+    def config(self) -> dict:
+        return {**super().config, "max_chars": self.max_chars}
+
+    def _run(self, data_registry: QaDataRegistry) -> Result:
+        v = getattr(getattr(data_registry, self.data), self.field)
+
+        if len(v) <= self.max_chars:
+            return self._result(passed=True)
+
+        return self._result(
+            passed=False,
+            message=self.failure_message,
+            offsets=[Offset(start=self.max_chars, end=len(v))],
+        )
+
+
+class NotTooShort(BaseGenericCheck):
+    name = "not_too_short"
+    display_name = "Not Too Short"
+    id = 10002
+    version = "1.0.0"
+    description = "The value meets or exceeds the minimum character length."
+    failure_message = "Text likely too short."
+
+    def __init__(
+        self,
+        *,
+        min_chars: int,
+        on_failure_policy: OnFailurePolicy,
+        data: str,
+        field: str,
+    ) -> None:
+        super().__init__(on_failure_policy=on_failure_policy, data=data, field=field)
+        self.min_chars = min_chars
+
+    @property
+    def config(self) -> dict:
+        return {**super().config, "min_chars": self.min_chars}
+
+    def _run(self, data_registry: QaDataRegistry) -> Result:
+        v = getattr(getattr(data_registry, self.data), self.field)
+
+        if len(v) >= self.min_chars:
+            return self._result(passed=True)
+
+        return self._result(
+            passed=False,
+            message=self.failure_message,
+            offsets=[Offset(start=0, end=len(v))],
+        )
