+++ qa/qa/checks/generic/punctuation.py
@@ -1,0 +1,36 @@
+"""Generic punctuation checks."""
+
+from qa.checks.base import BaseGenericPatternCheck
+
+
+class DoesNotEndWithPunctuation(BaseGenericPatternCheck):
+    name = "does_not_end_with_punctuation"
+    display_name = "Does Not End With Punctuation"
+    id = 10029
+    version = "1.0.0"
+    description = "The value does not end with punctuation."
+    failure_message = "Ends with punctuation."
+
+    _pattern = r"(?i)[!$%^&(_=`:;,.?-]$"
+
+
+class DoesNotEndWithPeriod(BaseGenericPatternCheck):
+    name = "does_not_end_with_period"
+    display_name = "Does Not End With Period"
+    id = 10058
+    version = "1.0.0"
+    description = "The value does not end with a period."
+    failure_message = "Ends with a period."
+
+    _pattern = r"\.$"
+
+
+class DoesNotContainSemicolon(BaseGenericPatternCheck):  # TODO remove?
+    name = "does_not_contain_semicolon"
+    display_name = "Does Not Contain Semicolon"
+    id = 10022
+    version = "1.0.0"
+    description = "The value does not contain a semicolon."
+    failure_message = "Contains semicolon(s) - use ',' or 'and' to separate authors."
+
+    _pattern = r";"
